+++ src/config.ts
@@ -22,8 +22,8 @@ const fileNames = [
 	'src/main.css',
 	'src/App.ts',
 	'src/widgets/HelloWorld.ts',
-	'src/widgets/styles/HelloWorld.css',
-	'src/widgets/styles/HelloWorld.css.d.ts',
+	'src/widgets/styles/HelloWorld.m.css',
+	'src/widgets/styles/HelloWorld.m.css.d.ts',
 	'tests/unit/all.ts',
 	'tests/unit/main.ts',
 	'tests/unit/App.ts',
