+++ src/templates/src/widgets/styles/HelloWorld.m.css.d.ts
